refactor: modularize diagnostics and structures
